+++ src/main/java/org/hspconsortium/sandboxmanager/model/Persona.java
@@ -9,19 +9,19 @@ import java.math.BigInteger;
                 query="SELECT c FROM Persona c WHERE c.fhirId = :id")
 })
 public class Persona {
-    private BigInteger id;
+    private Integer id;
     private String name;
     private String fhirId;
     private String resource;
     private String fullUrl;
 
-    public void setId(BigInteger id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
     @Id // @Id indicates that this it a unique primary key
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public BigInteger getId() {
+    public Integer getId() {
         return id;
     }
